+++ components/controls/ActionsDropdown.tsx
@@ -26,6 +26,9 @@ import SettingsIcon from '../icons/SettingsIcon';
 import { PopupPlacement } from '../ui/popups';
 import { Button } from './buttons';
 import Dropdown from './Dropdown';
+import { ClipboardBase } from './clipboard';
+import Tooltip from './Tooltip';
+import { translate } from '../../helpers/l10n';
 
 export interface ActionsDropdownProps {
   className?: string;
@@ -58,6 +61,8 @@ export default function ActionsDropdown(props: ActionsDropdownProps) {
 interface ItemProps {
   className?: string;
   children: React.ReactNode;
+  /** used to pass a string to copy to clipboard */
+  copyValue?: string;
   destructive?: boolean;
   /** used to pass a name of downloaded file */
   download?: string;
@@ -99,6 +104,22 @@ export class ActionsDropdownItem extends React.PureComponent<ItemProps> {
             {this.props.children}
           </Link>
         </li>
+      );
+    }
+
+    if (this.props.copyValue) {
+      return (
+        <ClipboardBase>
+          {({ setCopyButton, copySuccess }) => (
+            <Tooltip overlay={translate('copied_action')} visible={copySuccess}>
+              <li data-clipboard-text={this.props.copyValue} ref={setCopyButton}>
+                <a className={className} href="#" id={this.props.id} onClick={this.handleClick}>
+                  {this.props.children}
+                </a>
+              </li>
+            </Tooltip>
+          )}
+        </ClipboardBase>
       );
     }