+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - SONAR-12860 Add boolean and numberic type for RadioToggle and fix size glitch.
+- SC-1258 Add 'copy to clipboard' feature to ActionsDropdownItem
+- SC-1258 Add new popup positioning TopLeft
 
 ## 0.0.48